Promote network security mirroring resources to GA. (#13322)
[upstream:6510bb363bd4ccd11d64b42c11c444fb0ddfd5cb]
