+++ google-beta/services/networksecurity/resource_network_security_mirroring_deployment_group_generated_test.go
@@ -1,17 +1,36 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
+
+// ----------------------------------------------------------------------------
+//
+//     ***     AUTO GENERATED CODE    ***    Type: MMv1     ***
+//
+// ----------------------------------------------------------------------------
+//
+//     This file is automatically generated by Magic Modules and manual
+//     changes will be clobbered when the file is regenerated.
+//
+//     Please read more about how to change this file in
+//     .github/CONTRIBUTING.md.
+//
+// ----------------------------------------------------------------------------
+
 package networksecurity_test
 
 import (
+	"fmt"
+	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	"github.com/hashicorp/terraform-plugin-testing/plancheck"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
 
 	"github.com/hashicorp/terraform-provider-google-beta/google-beta/acctest"
+	"github.com/hashicorp/terraform-provider-google-beta/google-beta/tpgresource"
+	transport_tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta/transport"
 )
 
-func TestAccNetworkSecurityMirroringDeploymentGroup_update(t *testing.T) {
+func TestAccNetworkSecurityMirroringDeploymentGroup_networkSecurityMirroringDeploymentGroupBasicExample(t *testing.T) {
 	t.Parallel()
 
 	context := map[string]interface{}{
@@ -20,49 +39,34 @@ func TestAccNetworkSecurityMirroringDeploymentGroup_update(t *testing.T) {
 
 	acctest.VcrTest(t, resource.TestCase{
 		PreCheck:                 func() { acctest.AccTestPreCheck(t) },
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderBetaFactories(t),
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories(t),
+		CheckDestroy:             testAccCheckNetworkSecurityMirroringDeploymentGroupDestroyProducer(t),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccNetworkSecurityMirroringDeploymentGroup_basic(context),
+				Config: testAccNetworkSecurityMirroringDeploymentGroup_networkSecurityMirroringDeploymentGroupBasicExample(context),
 			},
 			{
 				ResourceName:            "google_network_security_mirroring_deployment_group.default",
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"labels", "terraform_labels"},
-			},
-			{
-				Config: testAccNetworkSecurityMirroringDeploymentGroup_update(context),
-				ConfigPlanChecks: resource.ConfigPlanChecks{
-					PreApply: []plancheck.PlanCheck{
-						plancheck.ExpectResourceAction("google_network_security_mirroring_deployment_group.default", plancheck.ResourceActionUpdate),
-					},
-				},
-			},
-			{
-				ResourceName:            "google_network_security_mirroring_deployment_group.default",
-				ImportState:             true,
-				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"update_time", "labels", "terraform_labels"},
+				ImportStateVerifyIgnore: []string{"labels", "location", "mirroring_deployment_group_id", "terraform_labels"},
 			},
 		},
 	})
 }
 
-func testAccNetworkSecurityMirroringDeploymentGroup_basic(context map[string]interface{}) string {
+func testAccNetworkSecurityMirroringDeploymentGroup_networkSecurityMirroringDeploymentGroupBasicExample(context map[string]interface{}) string {
 	return acctest.Nprintf(`
 resource "google_compute_network" "network" {
-  provider                = google-beta
   name                    = "tf-test-example-network%{random_suffix}"
   auto_create_subnetworks = false
 }
 
 resource "google_network_security_mirroring_deployment_group" "default" {
-  provider                      = google-beta
   mirroring_deployment_group_id = "tf-test-example-dg%{random_suffix}"
   location                      = "global"
   network                       = google_compute_network.network.id
-  description                   = "initial description"
+  description                   = "some description"
   labels = {
     foo = "bar"
   }
@@ -70,23 +74,41 @@ resource "google_network_security_mirroring_deployment_group" "default" {
 `, context)
 }
 
-func testAccNetworkSecurityMirroringDeploymentGroup_update(context map[string]interface{}) string {
-	return acctest.Nprintf(`
-resource "google_compute_network" "network" {
-  provider                = google-beta
-  name                    = "tf-test-example-network%{random_suffix}"
-  auto_create_subnetworks = false
-}
+func testAccCheckNetworkSecurityMirroringDeploymentGroupDestroyProducer(t *testing.T) func(s *terraform.State) error {
+	return func(s *terraform.State) error {
+		for name, rs := range s.RootModule().Resources {
+			if rs.Type != "google_network_security_mirroring_deployment_group" {
+				continue
+			}
+			if strings.HasPrefix(name, "data.") {
+				continue
+			}
 
-resource "google_network_security_mirroring_deployment_group" "default" {
-  provider                      = google-beta
-  mirroring_deployment_group_id = "tf-test-example-dg%{random_suffix}"
-  location                      = "global"
-  network                       = google_compute_network.network.id
-  description                   = "updated description"
-  labels = {
-    foo = "goo"
-  }
-}
-`, context)
+			config := acctest.GoogleProviderConfig(t)
+
+			url, err := tpgresource.ReplaceVarsForTest(config, rs, "{{NetworkSecurityBasePath}}projects/{{project}}/locations/{{location}}/mirroringDeploymentGroups/{{mirroring_deployment_group_id}}")
+			if err != nil {
+				return err
+			}
+
+			billingProject := ""
+
+			if config.BillingProject != "" {
+				billingProject = config.BillingProject
+			}
+
+			_, err = transport_tpg.SendRequest(transport_tpg.SendRequestOptions{
+				Config:    config,
+				Method:    "GET",
+				Project:   billingProject,
+				RawURL:    url,
+				UserAgent: config.UserAgent,
+			})
+			if err == nil {
+				return fmt.Errorf("NetworkSecurityMirroringDeploymentGroup still exists at %s", url)
+			}
+		}
+
+		return nil
+	}
 }
