+++ website/docs/r/network_security_mirroring_deployment.html.markdown
@@ -27,12 +27,10 @@ GENEVE-encapsulated replica traffic, e.g. a zonal instance group fronted by
 an internal passthrough load balancer. Deployments are always part of a
 global deployment group which represents a global mirroring service.
 
-~> **Warning:** This resource is in beta, and should be used with the terraform-provider-google-beta provider.
-See [Provider Versions](https://terraform.io/docs/providers/google/guides/provider_versions.html) for more details on beta resources.
 
 To get more information about MirroringDeployment, see:
 
-* [API documentation](https://cloud.google.com/network-security-integration/docs/reference/rest/v1beta1/projects.locations.mirroringDeployments)
+* [API documentation](https://cloud.google.com/network-security-integration/docs/reference/rest/v1/projects.locations.mirroringDeployments)
 * How-to Guides
     * [Mirroring deployment overview](https://cloud.google.com/network-security-integration/docs/out-of-band/deployments-overview)
 
@@ -46,13 +44,11 @@ To get more information about MirroringDeployment, see:
 
 ```hcl
 resource "google_compute_network" "network" {
-  provider                = google-beta
   name                    = "example-network"
   auto_create_subnetworks = false
 }
 
 resource "google_compute_subnetwork" "subnetwork" {
-  provider      = google-beta
   name          = "example-subnet"
   region        = "us-central1"
   ip_cidr_range = "10.1.0.0/16"
@@ -60,7 +56,6 @@ resource "google_compute_subnetwork" "subnetwork" {
 }
 
 resource "google_compute_region_health_check" "health_check" {
-  provider = google-beta
   name     = "example-hc"
   region   = "us-central1"
   http_health_check {
@@ -69,7 +64,6 @@ resource "google_compute_region_health_check" "health_check" {
 }
 
 resource "google_compute_region_backend_service" "backend_service" {
-  provider              = google-beta
   name                  = "example-bs"
   region                = "us-central1"
   health_checks         = [google_compute_region_health_check.health_check.id]
@@ -78,7 +72,6 @@ resource "google_compute_region_backend_service" "backend_service" {
 }
 
 resource "google_compute_forwarding_rule" "forwarding_rule" {
-  provider               = google-beta
   name                   = "example-fwr"
   region                 = "us-central1"
   network                = google_compute_network.network.name
@@ -91,14 +84,12 @@ resource "google_compute_forwarding_rule" "forwarding_rule" {
 }
 
 resource "google_network_security_mirroring_deployment_group" "deployment_group" {
-  provider                      = google-beta
   mirroring_deployment_group_id = "example-dg"
   location                      = "global"
   network                       = google_compute_network.network.id
 }
 
 resource "google_network_security_mirroring_deployment" "default" {
-  provider                   = google-beta
   mirroring_deployment_id    = "example-deployment"
   location                   = "us-central1-a"
   forwarding_rule            = google_compute_forwarding_rule.forwarding_rule.id
